Simplify FTS search service and contracts
